refactor: 질문  enum 및 ID 타입 일관성 정리 #32
