--- apps/be/src/common/enums/learning-category.enum.ts
@@ -1,5 +1,0 @@
-export enum LearningCategory {
-  OS = 'OS',
-  NETWORK = 'NETWORK',
-  DB = 'DB',
-}
