--- apps/be/src/common/enums/learning-difficulty.enum.ts
@@ -1,5 +1,0 @@
-export enum LearningDifficulty {
-  EASY = 'EASY',
-  MEDIUM = 'MEDIUM',
-  HARD = 'HARD',
-}
