+++ apps/be/src/modules/question-provider/domain/models/question.model.ts
@@ -1,8 +1,8 @@
-import { Difficulty, Domain } from '../../presentation/dto/pick-question.request.dto';
+import { Difficulty, Domain } from '@/common/enums/learning.enum';
 
 // 도메인 모델: 서비스/전략/레포 계층에서 공용으로 사용하는 질문 엔터티 모델
 export interface QuestionModel {
-  id: bigint;
+  id: number;
   domain: Domain;
   difficulty: Difficulty;
   topicId: string;
