+++ apps/be/src/modules/question-provider/domain/strategy/question-pick-strategy.ts
@@ -1,8 +1,8 @@
-import { Difficulty, Domain } from '../../presentation/dto/pick-question.request.dto';
+import { Difficulty, Domain } from '@/common/enums/learning.enum';
 
 // PickOptions: 확장 포인트 (세션/유저 이력은 지금 미구현)
 export interface PickOptions {
-  excludeQuestionIds?: bigint[];
+  excludeQuestionIds?: number[];
   userId?: string;
   sessionId?: string;
   strategy?: 'random' | 'avoid_seen';
@@ -10,7 +10,7 @@ export interface PickOptions {
 
 // 전략 패턴 인터페이스: 전략마다 질문 ID 선택 방법을 바꿀 수 있음
 export interface QuestionPickStrategy {
-  pick(domain: Domain, difficulty: Difficulty, options?: PickOptions): Promise<bigint>;
+  pick(domain: Domain, difficulty: Difficulty, options?: PickOptions): Promise<number>;
 }
 
 // DI 토큰
