+++ apps/be/src/modules/question-provider/domain/strategy/random-pick-strategy.ts
@@ -1,8 +1,9 @@
 import { Inject, Injectable } from '@nestjs/common';
 
+import { Difficulty, Domain } from '@/common/enums/learning.enum';
+
 import { QUESTION_POOL_CACHE } from '../../infra/ports/question-pool.cache';
 import type { QuestionPoolCachePort } from '../../infra/ports/question-pool.cache';
-import { Difficulty, Domain } from '../../presentation/dto/pick-question.request.dto';
 import { PickOptions, QuestionPickStrategy } from './question-pick-strategy';
 
 // 무작위 선택 전략: Redis SET에서 SRANDMEMBER 1개를 가져오는 기본 전략
@@ -10,7 +11,7 @@ import { PickOptions, QuestionPickStrategy } from './question-pick-strategy';
 export class RandomPickStrategy implements QuestionPickStrategy {
   constructor(@Inject(QUESTION_POOL_CACHE) private readonly pool: QuestionPoolCachePort) {}
 
-  async pick(domain: Domain, difficulty: Difficulty, _options?: PickOptions): Promise<bigint> {
+  async pick(domain: Domain, difficulty: Difficulty, _options?: PickOptions): Promise<number> {
     const id = await this.pool.getRandomId(domain, difficulty);
     if (!id) {
       throw new Error('Question pool not warmed; run warm-up');
