+++ apps/be/src/modules/question-provider/infra/prisma/question.repository.prisma.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@nestjs/common';
 
+import { Difficulty, Domain } from '@/common/enums/learning.enum';
+
 import { QuestionModel } from '../../domain/models/question.model';
-import { Difficulty, Domain } from '../../presentation/dto/pick-question.request.dto';
 import { QuestionRepositoryPort } from '../ports/question.repository';
 import { PrismaService } from './prisma.service';
 
@@ -10,14 +11,14 @@ import { PrismaService } from './prisma.service';
 export class QuestionRepositoryPrisma implements QuestionRepositoryPort {
   constructor(private readonly prisma: PrismaService) {}
 
-  async findById(id: bigint): Promise<QuestionModel | null> {
+  async findById(id: number): Promise<QuestionModel | null> {
     // Prisma schema에 맞춘 타입 캐스팅 필요 (@prisma/client 설치 후 조정)
     const row = await (this.prisma as any).question.findUnique({
       where: { id },
     });
     if (!row) return null;
     return {
-      id: BigInt(row.id),
+      id: row.id,
       domain: row.category as Domain,
       difficulty: row.difficulty as Difficulty,
       topicId: row.topicId,
@@ -28,11 +29,11 @@ export class QuestionRepositoryPrisma implements QuestionRepositoryPort {
     } satisfies QuestionModel;
   }
 
-  async findIdsByDomainDifficulty(domain: Domain, difficulty: Difficulty): Promise<bigint[]> {
+  async findIdsByDomainDifficulty(domain: Domain, difficulty: Difficulty): Promise<number[]> {
     const rows = await (this.prisma as any).question.findMany({
       where: { category: domain, difficulty },
       select: { id: true },
     });
-    return rows.map((r: any) => BigInt(r.id));
+    return rows.map((r: any) => r.id);
   }
 }
