+++ apps/be/src/modules/question-provider/presentation/dto/pick-question.response.dto.ts
@@ -1,8 +1,8 @@
-import { Difficulty, Domain } from './pick-question.request.dto';
+import { Difficulty, Domain } from '@/common/enums/learning.enum';
 
 // 응답 DTO: guide는 포함하지 않음 (AGENT.md 규칙)
 export class PickQuestionResponseDto {
-  questionId!: bigint | string;
+  questionId!: number | string;
   domain!: Domain;
   difficulty!: Difficulty;
   topicId!: string;
